Finishing Chapter 3 Implementation.
Checking in Chapter 3 implementation with the bonus problems done.  However, since we aren't serializing the question bank with answer status, every time we rotate, we aren't saving the state.  So the quiz needs to be done in one rotation pattern, which is kind of sucky, but I don't think I really care right now.